feat(add-new-chains): seperate chain type
RE #KRYS-2734
